% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubecaption-package.R
\docType{package}
\name{youtubecaption-package}
\alias{youtubecaption}
\alias{youtubecaption-package}
\title{youtubecaption: Downloading YouTube Subtitle Transcription in a Tidy Tibble Data_Frame}
\description{
Although there exist some R packages tailored for YouTube API (e.g., 'tuber'), downloading YouTube video subtitle (i.e., caption) in a tidy form has never been a low-hanging fruit. Using 'youtube-transcript-api Python package' under the hood, this R package provides users with a convenient way of parsing and converting a desired YouTube caption into a handy 'tibble' data_frame object. Furthermore, users can easily save a desired YouTube caption data as a tidy Excel file without advanced programming background knowledge.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/youtubecaption}
  \item Report bugs at \url{https://github.com/jooyoungseo/youtubecaption/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu} (\href{https://orcid.org/0000-0002-4064-6012}{ORCID})

Authors:
\itemize{
  \item Soyoung Choi \email{sxc940@psu.edu} (\href{https://orcid.org/0000-0002-0998-3352}{ORCID})
}

}
\keyword{internal}
